+++ CHANGELOG.md
@@ -29,6 +29,10 @@
   binaries.
 - Added the `tx_malformed_frames` metric for the virtio net device, emitted
   when a TX frame missing the VNET header is encountered.
+- Added metrics for counting rate limiter throttling events.
+- Added metric for counting MAC address updates.
+- Added metrics for counting TAP read and write errors.
+- Added metrics for counting RX and TX partial writes.
 
 ### Fixed